refactor: rename a module
